Pass ApexJobServiceTest.isExecuteBatchReady
The `ApexJobService` was refactored to return a singleton instance, so that
regardless of the type of job being submitted the service would know whether
a processing job had already been executed.

To work around the testing limitation where only a max of one batch may be
resolved by `Test.stopTest()`, the `submitAndExecuteTwice` test was replaced
with a simple check for the criteria that must be met in order for
`ApexJobService.executeBatch` to be called.