Pass SlashclockCallbackControllerTest.loadSuccess
Addressed technical debt where controller action method was handling
too much business logic. The business logic was moved into a new
`SlackService.activateTeam` method. Also, to facilitate testing the
previously-thought-to-be-clever personalization of `SlackService` and
`SlackApiService` was expanded to allow for generic instances to
process the same operations but with an added, explicitly needed Team
ID or access token.